+++ nds-service-client-impl/src/main/java/br/com/abril/nds/service/impl/ProdutoEdicaoServiceImpl.java
@@ -1726,7 +1726,7 @@ public class ProdutoEdicaoServiceImpl implements ProdutoEdicaoService {
     	for (EdicoesProdutosDTO edicoesProdutosDTO : historicoEdicoes) {
 			if(edicoesProdutosDTO.getStatus().equalsIgnoreCase("FECHADO") 
 					&& (edicoesProdutosDTO.getVenda() == null || edicoesProdutosDTO.getVenda().compareTo(BigInteger.ZERO)<=0)){
-				edicoesProdutosDTO.setVenda(estoqueProdutoService.obterVendaBaseadoNoEstoque(edicoesProdutosDTO.getIdProdutoEdicao()));
+				edicoesProdutosDTO.setVenda(estoqueProdutoService.obterVendaBaseadoNoEstoque(edicoesProdutosDTO.getIdProdutoEdicao().longValue()));
 			}
 		}
